Iniciando Policies Beneficiario e Conta

+++ app/Http/Controllers/BeneficiarioController.php
@@ -8,11 +8,13 @@ use Illuminate\Support\Facades\DB;
 class BeneficiarioController extends Controller
 {
     public function inicio(){
+        $this->authorize("create", \App\Models\Beneficiario::class);
 		$datas = DB::select("SELECT * FROM users WHERE tipo_usuario = '1'");
 		return view("adicionarbeneficiario", ['datas' => $datas]);
 	}
 	
 	public function adicionar($id){
+	    $this->authorize("create", \App\Models\Beneficiario::class);
 		$beneficiario = new Beneficiario();
 		$beneficiario->user_id = $id;
 		$beneficiario->save();
@@ -25,6 +27,7 @@ class BeneficiarioController extends Controller
     }
     
     public function listar(){
+        $this->authorize("view", \App\Models\Beneficiario::class);
         $beneficiarios = DB::select("SELECT * FROM beneficiarios WHERE deleted_at IS NULL");
         $datas = array();
         foreach($beneficiarios as $beneficiario){
@@ -35,6 +38,7 @@ class BeneficiarioController extends Controller
     }
     
     public function remover($id){
+        $this->authorize("delete", \App\Models\Beneficiario::class);
     	$beneficiario = Beneficiario::where('user_id', '=', $id)->first();
     	$beneficiario->delete();
     	return redirect("/beneficiarios/");
